refactor: MUI 테마 설정 위치 변경
