+++ .storybook/preview.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
-import type { Preview } from '@storybook/react';
 import { ThemeProvider } from '@mui/material';
+import type { Preview } from '@storybook/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import React from 'react';
 import '../src/styles/global.scss';
 
-import { MUI_THEME } from '../src/constants';
+import { MUI_THEME } from '../src/lib/mui/theme';
 
 const preview: Preview = {
   parameters: {
