+++ src/lib/mui/theme.ts
@@ -1,0 +1,16 @@
+import { colors, createTheme } from '@mui/material';
+
+export const MUI_THEME = createTheme({
+  palette: {
+    secondary: {
+      main: colors.blueGrey[800],
+    },
+  },
+  components: {
+    MuiButton: {
+      defaultProps: {
+        disableElevation: true,
+      },
+    },
+  },
+});
